updated the sop_app to a multipage app which shows predictor predictand maps, updated specs.cgi to work again and to show a link to the app
